requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
PyAudio==0.2.14
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
requests==2.31.0
srt==3.5.3
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.0
vosk==0.3.45
websockets==12.0
yarl==1.9.4
